[IMG-54] Move PixelValueType and PixelJustification enums to image class.
This is left-over scope from IMG-54.
